refactor: rename setModelValue to setValue
