+++ src/vue-wrapper.ts
@@ -197,7 +197,7 @@ export class VueWrapper<T extends ComponentPublicInstance> {
     return nextTick()
   }
 
-  setModelValue(value: any, prop?: string): Promise<void> {
+  setValue(value: any, prop?: string): Promise<void> {
     const propEvent = prop || 'modelValue'
     // @ts-ignore
     this.vm.$emit(`update:${propEvent}`, value)
